ingenious/document_parser/requirements.txt
annotated-types==0.7.0
black==25.1.0
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
coverage==7.9.1
cryptography==45.0.4
et-xmlfile==2.0.0
flake8==7.3.0
iniconfig==2.1.0
jinja2==3.1.6
lxml==6.0.0
markupsafe==3.0.2
mccabe==0.7.0
mypy==1.16.1
mypy-extensions==1.1.0
numpy==2.3.1
openpyxl==3.1.5
packaging==25.0
pandas==2.3.0
pathspec==0.12.1
pdfminer-six==20250506
pdfplumber==0.11.7
pillow==11.3.0
platformdirs==4.3.8
pluggy==1.6.0
pycodestyle==2.14.0
pycparser==2.22
pydantic==2.11.7
pydantic-core==2.33.2
pyflakes==3.4.0
pygments==2.19.2
pypdf2==3.0.1
pypdfium2==4.30.1
pytest==8.3.5
pytest-cov==6.1.1
python-dateutil==2.9.0.post0
python-docx==1.2.0
pytz==2025.2
ruff==0.11.10
six==1.17.0
typing-extensions==4.14.0
typing-inspection==0.4.1
tzdata==2025.2
xlrd==2.0.2
